Work around async_trait behavior
The `#[async_trait]` attribute/crate does a transformation to all async
methods, which as far as i can tell removes the `&self` lifetime, so our
logic to offset by 1 if `has_self == true` for both compared types has
an off-by-one-error here. This didn't error out earlier, since
`get_region_from_params` uses `Vec::get`, so "this index is out of
bounds" is just as `None` as "this generic param is not of kind
lifetime".

Also this is more a workaround than a fix. I'm not sure if we can do
something cleverer than "check if the last lifetim's name is
`'async_trait`".